% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cf.last.query}
\alias{cf.last.query}
\title{Retrieve Last Query Result from CliFlo}
\usage{
cf.last.query()
}
\description{
Retrieve the last query submitted to CliFlo instead of querying the database
again and losing subscription rows.
}
\details{
This function is a back up for when the clifro query has been submitted and
the data returned but has not been assigned. This saves the
user resubmitting queries and using more rows from their subscription than
needed.
}
\note{
Only the data from the last query is saved.
}

