% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cf.save.kml}
\alias{cf.save.kml}
\title{Save \code{cfStation} Locations to a KML File}
\usage{
cf.save.kml(station, file_name = "my_stations_", file_path = ".")
}
\arguments{
\item{station}{\code{cfStation} object containing one or more stations}

\item{file_name}{file name for the resulting KML file}

\item{file_path}{file path for the resulting KML file}
}
\description{
Save \code{cfStation} locations to a KML file
}
\details{
The \code{cf.save.kml} function is for \code{cfStation}
objects to allow for the spatial visualisation of the selected stations. The
resulting KML file is saved and can then be opened by programs like Google
Earth (TM). The resultant KML file has the station names and locations shown
with blue markers for open and red markers for closed stations. The agent
numbers, network ID's and date ranges are displayed as the descriptions.

If no file name is specified, unique names are produced in the current \R
working directory.
}
\note{
The .kml suffix is appended automatically if it isn't already present
in the \code{file_name} argument.
}
\examples{
\dontrun{
# A selection of four Auckland region stations down the East Coast to the
# upper Waitemata Harbour; Leigh 2 Ews, Warkworth Ews, Tiri Tiri Lighthouse
# and Henderson
my.stations = cf.station(17838, 1340, 1401, 12327)
my.stations

# Save these stations to a KML file
cf.save.kml(my.stations)

# Where is the subscription-free Reefton Ews station?
cf.save.kml(cf.station(), file_name = "reeftonEWS")

# Save all the Christchurch stations (using partial matching)
all.chch.st = cf.find.station("christ", status = "all", search = "region")
cf.save.kml(all.chch.st, file_name = "all_Chch_stations")
}
}
\seealso{
\code{\link{cf.station}} and \code{vignette("cfStation")} for
working with known stations, otherwise \code{\link{cf.find.station}} and
\code{vignette("choose-station")} for creating \code{cfStation} objects when
the agent numbers are unknown.
}

