% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cf_query}
\alias{cf_query}
\title{Retrieve Data from the National Climate Database}
\usage{
cf_query(user, datatype, station, start_date, end_date = now(tz),
  date_format = "ymd_h", tz = "Pacific/Auckland", quiet = FALSE)
}
\arguments{
\item{user}{a \code{\link{cfUser}} object.}

\item{datatype}{a \code{\link{cfDatatype}} object containing the datatypes to
be retrieved.}

\item{station}{a \code{\link{cfStation}} object containing the stations where
the datatypes will be retrieved from.}

\item{start_date}{a character, Date or POSIXt object indicating the start
date. If a character string is supplied the date format
should be in the form \code{yyyy-mm-dd-hh} unless
\code{date_format} is specified.}

\item{end_date}{a character, Date or POSIXt object indicating the start
date. If a character string is supplied the date format
is \code{yyyy-mm-dd-hh} unless \code{date_format} is
specified. Defaults to \code{\link[lubridate]{now}}.}

\item{date_format}{a character string matching one of \code{"ymd_h"},
\code{"mdy_h"}, \code{"ydm_h"} or \code{"dmy_h"}
representing the \code{\link[lubridate]{lubridate}}
date parsing function.}

\item{tz}{the timezone for which the start and end dates refer to. Conversion
to Pacific/Auckland time is done automatically through the
\code{\link[lubridate]{with_tz}} function. Defaults to
"Pacific/Auckland".}

\item{quiet}{logical. When \code{TRUE} the function evaluates without
             displaying customary messages. Messages from CliFlo are still
             displayed.}
}
\description{
Query the National Climate Database via CliFlo based on the \pkg{clifro} user
and selected datatypes, stations and dates.
}
\details{
The \code{cf_query} function is used to send the query built up from the
\pkg{clifro} user and the selected \pkg{clifro} datatypes and stations to the
National Climate Database via CliFlo.
}

