% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cf_station}
\alias{cfStation-class}
\alias{cf_station}
\title{Create a \pkg{clifro} Station Object}
\usage{
cf_station(...)
}
\arguments{
\item{...}{comma separated agent numbers}
}
\value{
\code{cfStation} object
}
\description{
Create a \code{cfStation} object containing the station information.
}
\details{
A \pkg{clifro} station object is created by the constructor function
\code{cf_station}. The unique agent numbers of the stations are all that is
required to create a \code{cfStation} object using the \code{cf_station}
function as the rest of the information is scraped from CliFlo.
}
\examples{
# Create a cfStation object for the Leigh 1 and 2 Ews stations
leigh.st = cf_station(1339, 1340)
leigh.st

# Add another column showing how long the station has been open for
leigh.df = as(leigh.st, "data.frame")
leigh.df$ndays = with(leigh.df, round(end - start))
leigh.df
}
\seealso{
\code{\link{cf_find_station}} for creating \code{cfStation} objects
when the agent numbers are not known and \code{vignette("cfStation")}
for working with clifro stations including spatial plotting in \R. For saving
\code{cfStation} objects as KML files refer to the vignette or
\code{\link{cf_save_kml}}.
}

