% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{cfUser-class}
\alias{cf.user}
\alias{cfUser}
\alias{cfUser-class}
\title{The Clifro User}
\usage{
cf.user(username = "public", password = character())
}
\arguments{
\item{username}{a character string to be used as the cliflo username}

\item{password}{a character string to be used as the cliflo password}
}
\description{
Allow the user to log into cliflo from \R to build their query.
}
\details{
An object inheriting from the cfUser class is created by the constructor
function \code{cf.user}. The user must have an active subscription to cliflo
in order to create a valid object, unless a 'public' user is sought.
Visit \url{http://cliflo.niwa.co.nz/} for more information and to subscribe
to cliflo.
}
\note{
For the 'public' user (see examples) only the Reefton Ews station data
is available.
}
\examples{
public.cfuser = cf.user(username = "public")
public.cfuser
}
\seealso{
\code{\link{valid_cfuser}} for details on the validation of cfUser
and \code{\link[clifro]{summary}} to summarise user information.
}

