% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{clifro}
\alias{clifro}
\alias{clifro-package}
\title{From CliFlo to \pkg{clifro}: Enhancing The National Climate Database With R}
\description{
Imports data from New Zealand's National Climate Database via CliFlo into \R
and provides data visualisation functions.
}
\details{
The \pkg{clifro} package is intended to simplify the process of data
extraction, formatting and visualisation from the CliFlo web portal. It
requires the user to build a query consisting of 3 main components; the user,
the datatype(s) and the station(s). These are
then combined using the \code{\link{cf_query}} function which sends the query
to the CliFlo database and returns the results that can easily be plotted
using the \code{\link[clifro]{plot}} function.

This package requires the user to already have a current subscription to the
National Climate Database unless a public user is sought where data is
limited to Reefton Ews. Subscription is free and can obtained from
\url{http://cliflo.niwa.co.nz/pls/niwp/wsubform.intro}.
}
\examples{
# Create a public user ----------------------------------------------------

public.user = cf_user()
public.user

# Select datatypes --------------------------------------------------------

# 9am Surface wind (m/s)
wind.dt = cf_datatype(2, 1, 4, 1)

# Daily Rain
rain.dt = cf_datatype(3, 1, 1)

# Daily temperature extremes
temp.dt = cf_datatype(4, 2, 1)

# Combine them together
all.dts = wind.dt + rain.dt + temp.dt
all.dts 

# Select the Reefton Ews station ------------------------------------------

reefton.st = cf_station()
reefton.st

# Submit the query --------------------------------------------------------

# Retrieve all data from ~ six months ago at 9am
reefton.data = cf_query(public.user, all.dts, reefton.st, paste(as.Date(Sys.time()) - 182, "9"))
reefton.data


# Plot the data -----------------------------------------------------------

plot(reefton.data, which = 1)
plot(reefton.data, which = 1, wind_plot = "speed")
plot(reefton.data, which = 1, wind_plot = "direction")
plot(reefton.data, which = 2)
plot(reefton.data, which = 2, include_runoff = FALSE)
plot(reefton.data, which = 3)
}
\keyword{package}

