% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cfDatatype.R, R/cfStation.R
\docType{methods}
\name{+,cfDatatype,cfDatatype-method}
\alias{+,cfDatatype,cfDatatype-method}
\alias{+,cfStation,cfStation-method}
\title{Arithmetic Operators for Clifro Objects}
\usage{
\S4method{+}{cfDatatype,cfDatatype}(e1, e2)

\S4method{+}{cfStation,cfStation}(e1, e2)
}
\arguments{
\item{e1}{a \code{cfDatatype} or \code{cfStation} object}

\item{e2}{an object matching the class of e1}
}
\description{
This operator allows you to add more datatypes or stations to
\code{cfDatatype} and \code{cfStation} objects respectively.
}

