% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{dimnames,dataFrame-method}
\alias{dimnames,dataFrame-method}
\alias{dim,dataFrame-method}
\title{Dimension Attributes of a Clifro Object}
\usage{
\S4method{dimnames}{dataFrame}(x)

\S4method{dim}{dataFrame}(x)
}
\arguments{
\item{x}{a \code{dataFrame} object

Specifically, a \code{dataFrame} object is any \code{\link{cfStation}} or 
\code{cfData} object. These functions are provided for the user to have (some)
familiar \code{data.frame}-type functions available for use on \pkg{clifro}
objects.}
}
\description{
Retrieve the dimensions or dimension names of a \code{dataFrame} object.
}
\seealso{
\code{\link{cf_query}} for creating \code{cfData} objects, and
  \code{\link{cf_station}} for creating \code{cfStation} objects.
}
