% Generated by roxygen2 (4.0.1): do not edit by hand
\name{direction_plot}
\alias{direction_plot}
\title{Plot Wind Direction Contours}
\usage{
plot(x, wind_plot = "direction", ggtheme = "grey", contours = 10,
n_col = 1, ...)
}
\arguments{
\item{x}{a cfData object containing wind data}

\item{wind_plot}{either \code{windrose}, \code{speed} or \code{direction}
depending on the type of wind plot}

\item{ggtheme}{character string indicating the \code{\link[ggplot2]{ggtheme}}
to be used for plotting.}

\item{contours}{the number of contour lines to draw.}

\item{n_col}{the number of columns of plots.}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}.}
}
\description{
Fit a 2D kernel estimated surface and plot the contours through time.
}
\details{
This is an optional
plot method for \code{cfData} objects which can be chosen by
specifying \code{wind_plot = 'direction'} in the \code{plot} function
call.
}

