% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cfDataList.R, R/cfDatatype.R, R/cfStation.R, R/dataFrame.R
\docType{methods}
\name{Extract}
\alias{$,dataFrame-method}
\alias{Extract}
\alias{[,cfDataList,ANY,ANY,ANY-method}
\alias{[,cfDatatype,ANY,missing,missing}
\alias{[,cfDatatype,ANY,missing,missing-method}
\alias{[,cfStation,ANY,ANY,ANY-method}
\alias{[,dataFrame,ANY,ANY,ANY-method}
\alias{[[,cfDataList-method}
\alias{[[,dataFrame-method}
\title{Subsetting Methods for Clifro Objects}
\usage{
\S4method{[}{cfDataList,ANY,ANY,ANY}(x, i, j)

\S4method{[[}{cfDataList}(x, i)

\S4method{[}{cfDatatype,ANY,missing,missing}(x, i, j, drop)

\S4method{[}{cfStation,ANY,ANY,ANY}(x, i, j, drop = TRUE)

\S4method{[[}{dataFrame}(x, i)

\S4method{[}{dataFrame,ANY,ANY,ANY}(x, i, j, drop)

\S4method{$}{dataFrame}(x, name)
}
\arguments{
\item{x}{a \pkg{clifro} object}

\item{i}{indices specifying elements to extract. Indices are
\code{numeric} or \code{character} vectors or empty (missing) or
\code{NULL}. Character vectors will be matched to the names of
the object.}

\item{j}{indices specifying elements to extract. Indices are
\code{numeric} or \code{character} vectors or empty (missing) or
\code{NULL}. Character vectors will be matched to the names of
the object.}

\item{drop}{if \code{TRUE}, the result is coerced to the lowest possible
            dimension. See \code{\link{drop}} for further details.}

\item{name}{a literal character string. This is partially matched to the
names of the object.}
}
\description{
Operators acting on \code{cfDataList}, \code{cfDatatype}, \code{cfStation},
and \code{dataFrame} objects.
}
\details{
These are methods for the generic operators for classes within \pkg{clifro}.
They are intended to give the user the familiar functionality of subsetting
\code{\link{data.frame}} objects.
}

