% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,cfData-method}
\title{Default \pkg{clifro} Plotting}
\usage{
plot(x, which, wind_plot = "windrose", ...)
}
\arguments{
\item{x}{a \code{cfData} or \code{cfDataList} object.}

\item{which}{a single number representing the dataframe in the list to plot.
Only used if \code{x} is a \code{cfDataList} object.}

\item{wind_plot}{if \code{x} contains wind data then either \code{windrose},
\code{speed} or \code{direction} determines the type of plot. See the table
below.}

\item{...}{arguments passed onto the different plotting methods.

This method is intended to simplify the data visualisation and exploration
of CliFlo data. The type of plot is determined by the type of the data output
from a \pkg{clifro} query. The following table links the datatypes to the
corresponding plot methods:

\tabular{ll}{
\strong{Datatype} \tab \strong{Method}\cr
Wind \tab \code{\link{windrose}} for windrose plots\cr
Wind \tab \code{\link{speed_plot}} for wind speed plots\cr
Wind \tab \code{\link{direction_plot}} for direction contour plots\cr
Rain \tab \code{\link{rain_plot}} for rainfall plots\cr
Screen Obs \tab \code{\link{screenobs_plot}} for dry and wet bulb
temperature plots\cr
Max/Min Temp \tab \code{\link{maxmin_plot}} for maximum, minimum and
average temperature plots\cr
Earth Temp \tab \code{\link{earthtemp_plot}} for earth temperature
plots\cr
Sunshine \tab \code{\link{sunshine_plot}} for accumulated sunshine hour
plots\cr
Pressure \tab \code{\link{pressure_plot}} for mean sea level atmospheric
pressure plots\cr
}}
}
\description{
Plot \pkg{clifro} data based on the datatype.
}
\seealso{
\code{\link{cf.query}} to retrieve the CliFlo data.
}

