% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pressure_plot}
\alias{pressure_plot}
\title{Plot Mean Sea Level Atmospheric Pressure}
\usage{
plot(x, ggtheme = "grey", free_y = FALSE, ...)
}
\arguments{
\item{x}{a cfData object containing pressure data.}

\item{ggtheme}{character string indicating the \code{\link[ggplot2]{ggtheme}}
to be used for plotting.}

\item{free_y}{logical value indicating if the range of the y axis should be
allowed to differ for each station.}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}.}
}
\description{
Plot the MSL atmospheric pressure through time.
}

