% Generated by roxygen2 (4.0.1): do not edit by hand
\name{speed_plot}
\alias{speed_plot}
\title{Plot Wind Speed Time Series}
\usage{
plot(x, wind_plot = "speed", ggtheme = "grey", free_y = FALSE, ...)
}
\arguments{
\item{x}{a cfData object containing wind data}

\item{wind_plot}{either \code{windrose}, \code{speed} or \code{direction}
depending on the type of wind plot}

\item{ggtheme}{character string indicating the \code{\link[ggplot2]{ggtheme}}
to be used for plotting.}

\item{free_y}{logical value indicating if the range of the y axis should be
allowed to differ for each station.}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}.}
}
\description{
Plot the wind speed through time, with a +/- standard
deviation interval.
}
\details{
This is an optional
plot method for \code{cfData} objects which can be chosen by
specifying \code{wind_plot = 'speed'} in the \code{plot} function call.
}

