% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfUser.R
\docType{methods}
\name{summary,cfUser-method}
\alias{summary,cfUser-method}
\title{Summarise User Information}
\usage{
\S4method{summary}{cfUser}(object)
}
\arguments{
\item{object}{an object of class \code{cfUser}.}
}
\description{
Show the subscription status for the \pkg{clifro} user
}

