% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sunshine_plot}
\alias{sunshine_plot}
\title{Plot Sunshine Hours}
\usage{
plot(x, ggtheme = "grey", free_y = FALSE, ...)
}
\arguments{
\item{x}{a cfData object containing sunshine data.}

\item{ggtheme}{character string indicating the \code{\link[ggplot2]{ggtheme}}
to be used for plotting.}

\item{free_y}{logical value indicating if the range of the y axis should be
allowed to differ for each station.}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}.}
}
\description{
Plot the duration of accumulated bright sunshine hours through time.
}

