% Generated by roxygen2 (4.0.1): do not edit by hand
\name{windrose}
\alias{windrose}
\title{Plot a windrose}
\usage{
windrose(speed, direction, facet, calm_wind = 0, variable_wind = 990,
  n_directions = 12, n_speeds = 5, spd_cuts, legend_title = "Wind Speed",
  col_pal = "GnBu", ggtheme = "grey", wind_summary = FALSE, ...)
}
\arguments{
\item{speed}{numeric vector of wind speeds.}

\item{direction}{numeric vector of wind directions.}

\item{facet}{facets used to plot the various windroses. Only one faceting
variable is currently allowed, see 'Details' below.}

\item{calm_wind}{the upper limit for wind speed that is considered calm.}

\item{variable_wind}{numeric code for variable winds (if applicable).}

\item{n_directions}{the number of direction bins to plot.}

\item{n_speeds}{the number of equally spaced wind speed bins to plot. This is
used if \code{spd_cuts} is missing.}

\item{spd_cuts}{numeric vector containing the cut points for the wind speed
intervals.}

\item{legend_title}{character string to be used for the legend title.}

\item{col_pal}{character string indicating the name of the
\code{\link[RColorBrewer]{RColorBrewer}} colour palette to be
used for plotting, see 'Recommendations' below.}

\item{ggtheme}{character string indicating the \code{\link[ggplot2]{ggtheme}}
to be used for plotting, see 'Recommendations' below.}

\item{wind_summary}{if \code{TRUE} a dataframe is returned containing a summary of
the wind speeds for each level of the facet.}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}.}
}
\value{
a \code{ggplot} object or a \code{\link{data.frame}} if
\code{wind_summary} is \code{TRUE}.
}
\description{
Plot a windrose showing the wind speed and direction for given sites.
}
\details{
This is the default plotting method for \code{cfData} objects but is also
intended to be used as a stand-alone function for any wind dataset. A
different windrose is plotted for each level of the faceting variable which
is coerced to a factor if necessary. The facets will generally be the station
where the data were collected, seasons or dates. Currently only one faceting
variable is allowed and is passed to \code{\link[ggplot2]{facet_wrap}} with
the formula \code{~facet}.

The wind summary dataframe returned if \code{wind_summary} is \code{TRUE}
contains the percentage of calm days (\code{speed <= calm_days}), percentage
of variable days (\code{speed == variable_days}) and quantiles calculated
from the empirical cumulative distribution functions for each facet.
}
\section{Colour Selection}{

For black and white windroses that may be preferred if plots are to be used
in journal articles for example, recommended \code{ggtheme}s are \code{'bw'},
\code{'linedraw'}, \code{'minimal'} or \code{'classic'} and
the \code{col_pal} should be \code{'Greys'}. Otherwise, any of the sequential
\code{\link[RColorBrewer]{RColorBrewer}} colour palettes are recommended for
colour plots.
}

